backend/requirements.txt
Django==4.2.13
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.2
django-cors-headers==4.4.0
psycopg2-binary==2.9.12
python-dotenv==1.0.1
pandas==2.2.3
openpyxl==3.1.2
pytest==8.3.2
pytest-django==4.9.0
factory-boy==3.3.1
